refactor: extract contracts, session_cache, turn_trace_collector to astra-turn-core (Phase 18)
Move 3 modules (~1,072 LOC) from runtime to astra-turn-core:
- contracts.rs: trait definitions (TurnSessionActivityWriter, TurnCoreEventWriter, etc.)
- session_cache.rs: cached assistant message building
- turn_trace_collector.rs: context assembly trace collection

Runtime originals replaced with re-exports. The observability integration
test stays in runtime (depends on runtime-only ObservabilityHub).

Turn-core: 1580 tests passing
Runtime: 3222 tests + 5 pre-existing remote skill failures
